Rolled Back to Release 8.1

+++ caffe2/core/timer.h
@@ -23,8 +23,7 @@ class Timer {
    */
   inline void Start() { start_time_ = clock::now(); }
   inline float NanoSeconds() {
-    return static_cast<float>(
-        std::chrono::duration_cast<ns>(clock::now() - start_time_).count());
+    return std::chrono::duration_cast<ns>(clock::now() - start_time_).count();
   }
   /**
    * @brief Returns the elapsed time in milliseconds.
